+++ aldrin/src/discoverer/test.rs
@@ -16,14 +16,14 @@ async fn any_object_no_services() {
 
     let obj = client.create_object(ObjectUuid::new_v4()).await.unwrap();
 
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
     let ev = discoverer.next_event().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
 
     obj.destroy().await.unwrap();
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
     let ev = discoverer.next_event().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Destroyed);
     assert_eq!(*ev.key(), 0);
@@ -31,7 +31,7 @@ async fn any_object_no_services() {
 
     discoverer.stop().await.unwrap();
     assert!(discoverer.next_event().await.is_none());
-    assert!(discoverer.finished());
+    assert!(discoverer.is_finished());
 
     client.join().await;
     broker.join().await;
@@ -52,14 +52,14 @@ async fn specific_object_no_services() {
         .await
         .unwrap();
 
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
     let ev = discoverer.next_event().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
     assert_eq!(ev.object_id(), obj.id());
 
     obj.destroy().await.unwrap();
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
     let ev = discoverer.next_event().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Destroyed);
     assert_eq!(*ev.key(), 0);
@@ -67,7 +67,7 @@ async fn specific_object_no_services() {
 
     discoverer.stop().await.unwrap();
     assert!(discoverer.next_event().await.is_none());
-    assert!(discoverer.finished());
+    assert!(discoverer.is_finished());
 
     client.join().await;
     broker.join().await;
@@ -89,7 +89,7 @@ async fn specific_object() {
         .await
         .unwrap();
 
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
     let ev = discoverer.next_event().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
@@ -98,7 +98,7 @@ async fn specific_object() {
     assert_eq!(ev.service_id(svc2.id().uuid), svc2.id());
 
     obj.destroy().await.unwrap();
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
     let ev = discoverer.next_event().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Destroyed);
     assert_eq!(*ev.key(), 0);
@@ -106,7 +106,7 @@ async fn specific_object() {
 
     discoverer.stop().await.unwrap();
     assert!(discoverer.next_event().await.is_none());
-    assert!(discoverer.finished());
+    assert!(discoverer.is_finished());
 
     client.join().await;
     broker.join().await;
@@ -128,7 +128,7 @@ async fn any_object() {
         .await
         .unwrap();
 
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
     let ev = discoverer.next_event().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
@@ -137,7 +137,7 @@ async fn any_object() {
     assert_eq!(ev.service_id(svc2.id().uuid), svc2.id());
 
     obj.destroy().await.unwrap();
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
     let ev = discoverer.next_event().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Destroyed);
     assert_eq!(*ev.key(), 0);
@@ -147,7 +147,7 @@ async fn any_object() {
     let svc1 = obj.create_service(svc1.id().uuid, 0).await.unwrap();
     let svc2 = obj.create_service(svc2.id().uuid, 0).await.unwrap();
 
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
     let ev = discoverer.next_event().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
@@ -156,7 +156,7 @@ async fn any_object() {
     assert_eq!(ev.service_id(svc2.id().uuid), svc2.id());
 
     obj.destroy().await.unwrap();
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
     let ev = discoverer.next_event().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Destroyed);
     assert_eq!(*ev.key(), 0);
@@ -164,7 +164,7 @@ async fn any_object() {
 
     discoverer.stop().await.unwrap();
     assert!(discoverer.next_event().await.is_none());
-    assert!(discoverer.finished());
+    assert!(discoverer.is_finished());
 
     client.join().await;
     broker.join().await;
@@ -177,7 +177,7 @@ async fn empty() {
 
     let mut discoverer = client.create_discoverer::<()>().build().await.unwrap();
     assert!(discoverer.next_event().await.is_none());
-    assert!(discoverer.finished());
+    assert!(discoverer.is_finished());
 
     client.join().await;
     broker.join().await;
@@ -203,7 +203,7 @@ async fn current_only() {
     let _svc1 = obj.create_service(ServiceUuid::new_v4(), 0).await.unwrap();
     let _svc2 = obj.create_service(ServiceUuid::new_v4(), 0).await.unwrap();
 
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
     let ev = discoverer.next_event().await.unwrap();
     assert_eq!(ev.kind(), DiscovererEventKind::Created);
     assert_eq!(*ev.key(), 0);
@@ -212,11 +212,11 @@ async fn current_only() {
     assert_eq!(ev.service_id(svc2.id().uuid), svc2.id());
 
     assert!(discoverer.next_event().await.is_none());
-    assert!(discoverer.finished());
+    assert!(discoverer.is_finished());
 
     discoverer.stop().await.unwrap();
     discoverer.start_current_only().await.unwrap();
-    assert!(!discoverer.finished());
+    assert!(!discoverer.is_finished());
 
     client.join().await;
     broker.join().await;